# Project-Mannager-System/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: PM Daily Work Management API
  version: 1.0.0
  description: |
    API cho ứng dụng quản lý công việc hằng ngày của quản lý dự án phần mềm.

    ## Quy ước chung
    - Base path: `/api/v1`
    - Xác thực: `Authorization: Bearer <accessToken>` (JWT HS256, hết hạn 15 phút). Refresh token (7 ngày) chỉ gửi ở `/auth/refresh` và `/auth/logout`.
    - Thời gian: ISO-8601 UTC (`timestamptz`). Ngày (date): `yyyy-MM-dd`.
    - Phân trang: `page` (≥ 0, mặc định 0), `size` (1–100, mặc định 20), `sort` (whitelist theo từng endpoint, VD `fullName,asc`).
    - Lỗi: format thống nhất (BR-GEN-05) — `timestamp, status, error, code, message, path, fieldErrors, traceId`. Catalog mã lỗi: `docs/design/05-error-handling-design.md`.
    - Dữ liệu đã xóa mềm: không xuất hiện ở danh sách; truy cập trực tiếp trả 404.
    - Cập nhật bắt buộc kèm `version`; version cũ → 409.
    - Chi tiết từng module: `docs/api/01..12-*.md`.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /risks:
    get:
      tags: [Risk]
      summary: Danh sách risk
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - { $ref: '#/components/parameters/Sort' }
        - { $ref: '#/components/parameters/ProjectIdQuery' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/RiskStatus' } }
        - { name: level, in: query, schema: { $ref: '#/components/schemas/RiskLevel' } }
        - { name: ownerId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Trang kết quả
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskPage' }
        '403': { $ref: '#/components/responses/Forbidden403' }
    post:
      tags: [Risk]
      summary: Tạo risk (mã tự sinh RSK000001, level tự tính)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RiskCreateRequest' }
      responses:
        '201':
          description: Đã tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /risks/{id}:
    get:
      tags: [Risk]
      summary: Chi tiết risk
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RiskIdPath' }
      responses:
        '200':
          description: Chi tiết
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
    put:
      tags: [Risk]
      summary: "Cập nhật risk (owner: chỉ status)"
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RiskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RiskUpdateRequest' }
      responses:
        '200':
          description: Đã cập nhật
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RiskResponse' }
        '400': { $ref: '#/components/responses/Validation400' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409': { $ref: '#/components/responses/Conflict409' }
    delete:
      tags: [Risk]
      summary: Xóa mềm risk
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RiskIdPath' }
      responses:
        '204': { description: Thành công, không body }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
  /risks/{id}/convert-to-issue:
    post:
      tags: [Risk]
      summary: Chuyển risk OCCURRED thành issue (liên kết 1–1)
      security: [{ bearerAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/RiskIdPath' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConvertToIssueRequest' }
      responses:
        '201':
          description: Issue được tạo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IssueResponse' }
        '400':
          description: BAD_REQUEST khi risk chưa OCCURRED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403': { $ref: '#/components/responses/Forbidden403' }
        '404': { $ref: '#/components/responses/NotFound404' }
        '409':
          description: ALREADY_LINKED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  # ============================= ISSUE =============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
      description: Số trang (từ 0)
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Số bản ghi mỗi trang
    Sort:
      name: sort
      in: query
      schema: { type: string }
      description: 'Sắp xếp theo whitelist của endpoint, VD `fullName,asc`'
...
    ProjectIdQuery:
      name: projectId
      in: query
      schema: { type: string, format: uuid }
      description: Lọc theo dự án
...
    RiskIdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: ID risk
...
  responses:
    Validation400:
      description: Dữ liệu không hợp lệ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Forbidden403:
      description: Không đủ quyền hoặc ngoài phạm vi dữ liệu
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound404:
      description: Không tìm thấy (bao gồm dữ liệu đã xóa mềm)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict409:
      description: Xung đột optimistic locking (version cũ)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [timestamp, status, error, code, message, path, fieldErrors, traceId]
      properties:
        timestamp: { type: string, format: date-time, example: '2026-08-01T10:00:00Z' }
        status: { type: integer, example: 400 }
        error: { type: string, example: BAD_REQUEST }
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: 'Dữ liệu không hợp lệ' }
        path: { type: string, example: /api/v1/tasks }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
        traceId: { type: string, example: a1b2c3d4-1234 }
    PageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages, hasNext, hasPrevious]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer }
        totalPages: { type: integer }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
    UserRef:
      type: object
      properties:
        id: { type: string, format: uuid }
        fullName: { type: string }
    # ---------- Enum ----------
...
    RiskStatus: { type: string, enum: [OPEN, MONITORING, MITIGATED, OCCURRED, CLOSED] }
    RiskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    RiskProbability: { type: string, enum: [LOW, MEDIUM, HIGH] }
    RiskImpact: { type: string, enum: [LOW, MEDIUM, HIGH] }
    IssueSeverity: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
    IssueStatus: { type: string, enum: [OPEN, ANALYZING, IN_PROGRESS, RESOLVED, CLOSED, REJECTED] }
...
    RiskCreateRequest:
      type: object
      required: [projectId, title, probability, impact, ownerId]
      properties:
        projectId: { type: string, format: uuid }
        title: { type: string, maxLength: 200 }
        description: { type: string }
        probability: { $ref: '#/components/schemas/RiskProbability' }
        impact: { $ref: '#/components/schemas/RiskImpact' }
        level:
          $ref: '#/components/schemas/RiskLevel'
          description: Tự tính theo xác suất × ảnh hưởng nếu bỏ trống
        ownerId: { type: string, format: uuid }
        mitigationPlan: { type: string }
        contingencyPlan: { type: string }
        status: { $ref: '#/components/schemas/RiskStatus' }
        dueDate: { type: string, format: date }
    RiskUpdateRequest:
      allOf:
        - { $ref: '#/components/schemas/RiskCreateRequest' }
        - type: object
          required: [version]
          properties:
            version: { type: integer }
    ConvertToIssueRequest:
      type: object
      properties:
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        dueDate: { type: string, format: date }
    RiskResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: RSK000001 }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        probability: { $ref: '#/components/schemas/RiskProbability' }
        impact: { $ref: '#/components/schemas/RiskImpact' }
        level: { $ref: '#/components/schemas/RiskLevel' }
        owner: { $ref: '#/components/schemas/UserRef' }
        mitigationPlan: { type: string }
        contingencyPlan: { type: string }
        status: { $ref: '#/components/schemas/RiskStatus' }
        dueDate: { type: string, format: date, nullable: true }
        linkedIssueId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        version: { type: integer }
    RiskPage:
      allOf:
        - { $ref: '#/components/schemas/PageResponse' }
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/RiskResponse' }
    # ---------- Issue ----------
...
    IssueResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: ISS000001 }
        projectId: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        severity: { $ref: '#/components/schemas/IssueSeverity' }
        owner: { $ref: '#/components/schemas/UserRef' }
        rootCause: { type: string }
        solution: { type: string }
        status: { $ref: '#/components/schemas/IssueStatus' }
        dueDate: { type: string, format: date, nullable: true }
        resolvedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        version: { type: integer }
...